add: apiServerとaxiosHeadersを追加
