+++ vue/src/mixin/auth.js
@@ -1,3 +1,5 @@
+const apiServer = "http://localhost:3000";
+
 const removeItem = () => {
   window.localStorage.removeItem("uid");
   window.localStorage.removeItem("access-token");
@@ -12,4 +14,26 @@ const setItem = (res) => {
   window.localStorage.setItem("nickname", res.data.data.nickname);
 };
 
-export { setItem, removeItem };
+const getItem = () => {
+  const uid = window.localStorage.getItem("uid");
+  const client = window.localStorage.getItem("client");
+  const accessToken = window.localStorage.getItem("access-token");
+  const nickname = window.localStorage.getItem("nickname");
+  return {
+    uid: uid,
+    client: client,
+    accessToken: accessToken,
+    nickname: nickname,
+  };
+};
+
+const axiosHeaders = () => {
+  const { uid, client, accessToken } = getItem();
+  return {
+    uid: uid,
+    client: client,
+    "access-token": accessToken,
+  };
+};
+
+export { setItem, removeItem, getItem, apiServer, axiosHeaders };
